feat: add folder structure and arg parser
