+++ package.json
@@ -8,7 +8,7 @@
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "bin": {
-    "miragejs-open-api": "index.js"
+    "miragejs-open-api": "./index.js"
   },
   "keywords": [],
   "author": "",
